# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DOGFOOD 2026 Hackathon Platform API",
    "version": "1.0.0",
    "description": "Comprehensive REST API specification for the DOGFOOD 2026 Hackathon Platform, covering authentication, team formation, project submissions, judging, community voting, audit logs, and cryptographic attestations."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local / Container development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "sessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "Django session cookie authentication (e.g. Cookie: session=...)"
      }
    },
    "schemas": {
...
      "JudgeScoreItem": {
        "type": "object",
        "properties": {
          "project_id": { "type": "integer" },
          "project_title": { "type": "string" },
          "track": { "type": "string" },
          "criteria_scores": { "type": "object", "additionalProperties": { "type": "number" } },
          "comment": { "type": "string" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      }
    }
  },
  "paths": {
...
    "/api/judge/scores/": {
      "get": {
        "summary": "List judge evaluations with strict role isolation",
        "tags": ["Judging"],
        "security": [{ "sessionAuth": [] }],
        "parameters": [
          {
            "name": "judge",
            "in": "query",
            "required": false,
            "description": "Target judge username or ID (peers blocked for role 'judge')",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "List of scored projects and rubrics",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/JudgeScoreItem" }
                }
              }
            }
          },
          "401": { "description": "Unauthenticated" },
          "403": { "description": "Unauthorized role or peer score access attempt" }
        }
      }
    },
    "/api/export.csv": {
      "get": {
        "summary": "Export project evaluations as RFC 4180 CSV",
        "tags": ["Judging"],
        "security": [{ "sessionAuth": [] }],
        "responses": {
          "200": {
            "description": "CSV attachment with headers: project_id,project_title,track,submitted_at,review_count",
            "content": { "text/csv": { "schema": { "type": "string" } } }
          },
          "401": { "description": "Unauthenticated" },
          "403": { "description": "Only organizer or admin roles allowed" }
        }
      }
    },
...
  }
}